requirements.txt
openai-agents
python-dotenv
streamlit
beautifulsoup4
requests